refactor to ovalutil
