+++ rhel/parser.go
@@ -1,4 +1,4 @@
-package suse
+package rhel
 
 import (
 	"context"
@@ -7,24 +7,21 @@ import (
 	"io"
 
 	"github.com/quay/claircore"
-	"github.com/quay/claircore/libvuln/driver"
 	"github.com/quay/claircore/pkg/ovalutil"
 	"github.com/quay/goval-parser/oval"
 	"github.com/rs/zerolog"
 )
 
-var _ driver.Parser = (*Updater)(nil)
-
 func (u *Updater) Parse(ctx context.Context, r io.ReadCloser) ([]*claircore.Vulnerability, error) {
 	log := zerolog.Ctx(ctx).With().
-		Str("component", "suse/Updater.Parse").
+		Str("component", "rhel/Updater.Parse").
 		Logger()
 	ctx = log.WithContext(ctx)
 	log.Info().Msg("starting parse")
 	defer r.Close()
 	root := oval.Root{}
 	if err := xml.NewDecoder(r).Decode(&root); err != nil {
-		return nil, fmt.Errorf("suse: unable to decode OVAL document: %w", err)
+		return nil, fmt.Errorf("rhel: unable to decode OVAL document: %w", err)
 	}
 	log.Debug().Msg("xml decoded")
 	protoVuln := func(def oval.Definition) (*claircore.Vulnerability, error) {
@@ -34,7 +31,7 @@ func (u *Updater) Parse(ctx context.Context, r io.ReadCloser) ([]*claircore.Vuln
 			Description: def.Description,
 			Links:       ovalutil.Links(def),
 			Severity:    def.Advisory.Severity,
-			// each updater is configured to parse a suse release
+			// each updater is configured to parse a rhel release
 			// specific xml database. we'll use the updater's release
 			// to map the parsed vulnerabilities
 			Dist: releaseToDist(u.release),
